Add files via upload
This project is a web-based application that allows you to search for images using keywords or image URLs. It is built with HTML, CSS, and JavaScript and uses the Bing Image Search API to fetch the results. You can also filter the results by size, color, type, and license. You can view the source code and a live demo of the project from this